c/float_asm: silence -Wdouble-promotion, again.
On Alpine, with musl and clang, the compiler reports:

    float_asm.c:289:5: error: implicit conversion increases
    floating-point precision: 'float' to 'double'
    [-Werror,-Wdouble-promotion]
        ok_nan(sqrt_f(-1));
        ^~~~~~~~~~~~~~~~~~
    float_asm.c:239:80: note: expanded from macro 'ok_nan'
    #define ok_nan(fctcall) _ok(#fctcall, isnanf(fctcall), "f", (double)(fctcall), NAN)
                            ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^~~~
    /usr/include/math.h:15:19: note: expanded from macro 'NAN'
    #define NAN       __builtin_nanf("")
                      ^~~~~~~~~~~~~~~~~~
    float_asm.c:207:92: note: expanded from macro '_ok'
                printf("[FAIL] %s = %" format ", expected %" format "\n", description, result, expected); \
                ~~~~~~                                                                         ^~~~~~~~

+++ c/float_asm.c
@@ -236,7 +236,7 @@ static double invsqrt_d(double x)
 #if !defined(isnanf) && !defined(__GLIBC__)
 #    define isnanf(value) isnan((float)(value))
 #endif
-#define ok_nan(fctcall) _ok(#fctcall, isnanf(fctcall), "f", (double)(fctcall), NAN)
+#define ok_nan(fctcall) _ok(#fctcall, isnanf(fctcall), "f", (double)(fctcall), (double)NAN)
 
 static int check_constants(void)
 {
